==> Codes/Step1_parse_data.ipynb <==
### Summary

We sourced our initial data from Los Angeles Open Data. At the time of this project, Yelp limited daily API calls to 5000. In order to capture data for each restaurant accurately, it was necessary to perform two calls per restaurant. Our first step, necessarily, was to parse the initial 48k row data set into chunks of 2500. The code for performing these calls is stored in the YelpPulls notebook. 

Please review yelp_pulls for our next step.

import pandas as pd


city = pd.read_csv("health_scores.csv")
city.head()

rest_only = city[city['pe_description'].str.contains("RESTAURANT") == True]
rest_only.head()

rest_only.value_counts()

max_rows = 2500
dataframes = []
while len(rest_only) > max_rows:
    top = rest_only[:max_rows]
    dataframes.append(top)
    rest_only = rest_only[max_rows:]
else:
    dataframes.append(rest_only)

for _, frame in enumerate(dataframes):
    frame.to_csv(str(_)+'la_restaurants.csv', index=False)
