==> Codes/Step3_raw_all_restaurants.ipynb <==
### Summary

Once we had run all of our raw data through the Yelp API, it was necessary to merge all the output files into one place. We did this by looping through the folder of saved data and concatinating each resulting DataFrame together. This prepared our data to be cleaned all at once. This code can be found in the clean_all_restaurants notebook.

#Magic
%matplotlib

#Dependencies
import os
import pandas as pd
import numpy as np
import glob
from matplotlib import pyplot as plt

filepath = "called/**"
all_csv_files = glob.glob(os.path.join(filepath, "*.csv"), recursive = True)

grades_n_yelp_df = (pd.read_csv(f) for f in all_csv_files)
raw_restaurants_df = pd.concat(grades_n_yelp_df, ignore_index=True)


#all_restaurants_df=pd.concat(map(pd.read_csv, filepath))


raw_restaurants_df.head()

# Output BIG dataframe to CSV
raw_restaurants_df.to_csv("raw_all_restaurants.csv", header=True)
