==> Codes/Step4_clean_all_restaurants.ipynb <==
### Summary

We cleaned up superflous columns containing data that was not necessary to perform our analyses. Additionally, the data is organized according to inspection record, so there were somtimes duplicate records for a restaurant. We kept records according to the most recent inspection date in order to preserve the most up to date health score. Our next step was to place our data in context, and we thought a map would be the best medium for doing so. Please refer to the map_scores_and_ratings notebook.

#Magic
%matplotlib

#Dependencies
import os
import pandas as pd
import numpy as np
import glob
from matplotlib import pyplot as plt

# Not every CSV has encoding, yet better safe than sorry
raw_restaurants_df = pd.read_csv("raw_all_restaurants.csv", encoding="ISO-8859-1")


raw_restaurants_df.head()

raw_restaurants_df.count()

print(raw_restaurants_df.columns)

raw_restaurants_df = raw_restaurants_df.sort_values(["facility_id","activity_date"], ascending=False)

#raw_restaurants_df.drop_duplicates(subset=["activity_date"], take_last=False)

raw_restaurants_df.head()

#NOTE TO TEAM: "Unnamed: 0", "Unnamed: 0.1","index", and "owner_id" will not drop--the latter cannot be found in the axis ... not sure how to drop
raw_restaurants_df.drop(["Unnamed: 0", "Unnamed: 0.1", "index", "owner_id","serial_number", "service_code", "service_description", "employee_id", "owner_name", "program_element_pe", "program_name", "program_status"], axis=1, inplace=True)

raw_restaurants_df.head()

raw_restaurants_df = raw_restaurants_df.rename(columns={"unnamed (index)": "Index #",
"activity_date": "Inspection Date",
"facility_name": "Restaurant",
"score": "Score",
"grade": "Grade",
"facility_id": "Facility ID",
"facility_address": "Address",
"facility_city": "City",
"facility_state": "State",
"facility_zip": "Zip", 
"index": "IGNORE",                                  
"owner_id": "Owner ID",
"pe_description": "Type",
"record_id": "Inspection ID",           
"food_type": "Cuisine",
"in_business": "Active Business",
"latitude": "Latitude",
"longitude": "Longitude", 
"price": "Price",
"rating": "Rating",
"review_count": "# of Reviews",
"yelp_id": "Yelp ID"
})

raw_restaurants_df.head()

raw_restaurants_df.drop_duplicates(raw_restaurants_df.columns.difference(["Inspection Date"]), keep="first", inplace=True)

raw_restaurants_df.head()

raw_restaurants_df.count()

raw_restaurants_df.groupby(["Facility ID","Restaurant"]).size()

unique_restaurants = len(raw_restaurants_df["Restaurant"].unique())

                         
unique_restaurants

#Do a count of all columns
raw_restaurants_df.count()

#Average Inspection Score
raw_restaurants_df["Score"].mean()

#Average Yelp Rating
raw_restaurants_df["Rating"].mean()

#@TEAM: I don't understand why the new column names don't work--can anyone explain to me? --charla
restaurant_numbers_df = raw_restaurants_df[["Restaurant", "Score", "Rating", "# of Reviews"]]
restaurant_numbers_df.head()

# Output BIG dataframe of clean data to CSV
raw_restaurants_df.to_csv("clean_data_all_restaurants.csv", header=True)
