==> Codes/Step5_map_scores_and_ratings.ipynb <==
### Summary

There were 52k records in our initial dataset, and we wanted a visual of those records on a map. We were curious to see if there were any trends in Health Grade or Yelp Score by neighborhood. No immediate trends were observed. The next step would be to survey a topline number, so we looked at the %to total breakdown of Yelp Rating by Health Score.

# Import dependencies
import gmaps
import numpy as np
import pandas as pd
import requests

# Import raw data (i.e., sample data: "..\Data (Raw)\Yelp API\la_restaurants_yelped5.csv")
raw_data = pd.read_csv("..\Data (Clean)\clean_data_all_restaurants.csv", low_memory=False)
list(raw_data)

# Configure gmaps with API key
gmaps.configure(api_key='API KEYS')

## Yelp Rating Heatmap

# Store 'Lat' and 'Lng' into  locations 
locations = raw_data[["Latitude", "Longitude"]].astype(float)

# Convert 'Rating' to float and store
rating = raw_data["Rating"].astype(float)

locations = locations.reset_index(drop=True)
rating = rating.reset_index(drop=True)
rating.head()

# Create rating Heatmap layer
fig = gmaps.figure()

heat_layer = gmaps.heatmap_layer(locations, weights=rating, 
                                 dissipating=False, max_intensity=50,
                                 point_radius = .005)

# Adjust heat_layer setting to help with heatmap dissipating on zoom
heat_layer.dissipating = False
heat_layer.max_intensity = 50
heat_layer.point_radius = .005

fig.add_layer(heat_layer)

fig

## Health Score Heatmap

# Convert 'Score' to float and store
score = raw_data["Score"].astype(float)

score = score.reset_index(drop=True)
score.head()

# Create rating Heatmap layer
fig = gmaps.figure()

heat_layer = gmaps.heatmap_layer(locations, weights=score, 
                                 dissipating=False, max_intensity=1000,
                                 point_radius = .005)

# Adjust heat_layer setting to help with heatmap dissipating on zoom
heat_layer.dissipating = False
heat_layer.max_intensity = 1000
heat_layer.point_radius = .005

fig.add_layer(heat_layer)

fig
